requirements.txt
requests>=2.26.0
beautifulsoup4>=4.9.3
discord.py>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=0.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
psutil>=5.9.0